// typechain-types/ExampleExternalContract.ts
/* Autogenerated file. Do not edit manually. */
/* tslint:disable */
/* eslint-disable */
import type {
  BaseContract,
  BytesLike,
  FunctionFragment,
  Result,
  Interface,
  ContractRunner,
  ContractMethod,
  Listener,
} from "ethers";
import type {
  TypedContractEvent,
  TypedDeferredTopicFilter,
  TypedEventLog,
  TypedListener,
  TypedContractMethod,
} from "./common";

export interface ExampleExternalContractInterface extends Interface {
  getFunction(
    nameOrSignature: "complete" | "completed" | "getbalance"
  ): FunctionFragment;

  encodeFunctionData(functionFragment: "complete", values?: undefined): string;
  encodeFunctionData(functionFragment: "completed", values?: undefined): string;
  encodeFunctionData(
    functionFragment: "getbalance",
    values?: undefined
  ): string;

  decodeFunctionResult(functionFragment: "complete", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "completed", data: BytesLike): Result;
  decodeFunctionResult(functionFragment: "getbalance", data: BytesLike): Result;
}

export interface ExampleExternalContract extends BaseContract {
  connect(runner?: ContractRunner | null): ExampleExternalContract;
  waitForDeployment(): Promise<this>;

  interface: ExampleExternalContractInterface;

  queryFilter<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;
  queryFilter<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    fromBlockOrBlockhash?: string | number | undefined,
    toBlock?: string | number | undefined
  ): Promise<Array<TypedEventLog<TCEvent>>>;

  on<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  on<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  once<TCEvent extends TypedContractEvent>(
    event: TCEvent,
    listener: TypedListener<TCEvent>
  ): Promise<this>;
  once<TCEvent extends TypedContractEvent>(
    filter: TypedDeferredTopicFilter<TCEvent>,
    listener: TypedListener<TCEvent>
  ): Promise<this>;

  listeners<TCEvent extends TypedContractEvent>(
    event: TCEvent
  ): Promise<Array<TypedListener<TCEvent>>>;
  listeners(eventName?: string): Promise<Array<Listener>>;
  removeAllListeners<TCEvent extends TypedContractEvent>(
    event?: TCEvent
  ): Promise<this>;

  complete: TypedContractMethod<[], [void], "payable">;

  completed: TypedContractMethod<[], [boolean], "view">;

  getbalance: TypedContractMethod<[], [bigint], "view">;

  getFunction<T extends ContractMethod = ContractMethod>(
    key: string | FunctionFragment
  ): T;

  getFunction(
    nameOrSignature: "complete"
  ): TypedContractMethod<[], [void], "payable">;
  getFunction(
    nameOrSignature: "completed"
  ): TypedContractMethod<[], [boolean], "view">;
  getFunction(
    nameOrSignature: "getbalance"
  ): TypedContractMethod<[], [bigint], "view">;

  filters: {};
}